split DocumentEditor/toolbar-state into shared/ui
